add initial support for codex32
* Refactor Bech32 to separate validation, decoding, and checksum

Also make the charset and some things public, and add a new flag
to convertBits whether to enforce the last padded bits are zeros.
This will be used to support Codex32/BIP93.

* Add initial BIP93 support

This adds initial support for Codex32, specifically support for
importing secret shares and deriving BIP32 secrets from them.
Deriving shares is not yet implemented.